增加价格和发货方式模块
